requirements.txt
Django==2.0.5
pytz==2018.4
django-filter==1.1.0
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
gunicorn==19.8.0
Markdown==2.6.11
psycopg2==2.7.4
PyJWT==1.6.1
